Merged PR 3965: User Agent Client Hints Support
FEAT: The SetHeadersElement can be used to convert values from all 'SetHeader*' properties into a dictionary that can be used to populate the HTTP response headers. This element is added automatically for web integration pipelines.
FEAT: When using web integration, the pipeline will now automatically set HTTP headers in the response based on the values of properties starting with 'SetHeader'
TEST: Added tests for SetHeadersElement.